backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
numpy
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
openai>=1.0.0
google-generativeai>=0.3.0
gpt4all>=2.0.0
pytest
requests
python-multipart
tf-keras
torch
python-dotenv
